+++ doc/changelog.rst
@@ -1,6 +1,7 @@
 Changelog
 =========
 
+* :feature:`92` Use journald logging driver when available during docker execution
 * :feature:`88` Add option to specify more compose file to ``cob docker run-image``
 * :feature:`89` Add ``--force`` to ``cob docker deploy`` to force overwriting unit files
 * :feature:`90` Add ``docker.exposed_ports`` configuration for controlling exposed ports in deployment